Add predictor
The algorithm of predictor:
Track local (in one time unit) and global frequency history.
For each prediction of victim bits, set the threshold and test local
first, if over the threshold then set the victim bit to true. Second,
compare the global history and do same thing. In addition, set the
size threshold that if the number of bits which are set is greater
than size then return. After local and global setting, if there are
other additional spaces then set those global history is greater than
0. Finally, setting the victim bits from small index to big.